requirements.txt
fastapi==0.95.1
google-cloud-datastore==2.15.2
gunicorn==20.1.0
uvicorn==0.22.0
